[FLINK-29909][Connectors/DynamoDB] Update package names to follow convention of other AWS connectors
