--- flink-connector-dynamodb/src/main/java/org/apache/flink/streaming/connectors/dynamodb/sink/DynamoDbSinkException.java
@@ -1,52 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.streaming.connectors.dynamodb.sink;
-
-import org.apache.flink.annotation.PublicEvolving;
-
-/** Exception is thrown when DynamoDb sink failed to write data. */
-@PublicEvolving
-public class DynamoDbSinkException extends RuntimeException {
-    public DynamoDbSinkException(String message) {
-        super(message);
-    }
-
-    public DynamoDbSinkException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    /**
-     * When the flag {@code failOnError} is set in {@link DynamoDbSinkWriter}, this exception is
-     * raised as soon as any non-recoverable exception occurs.
-     */
-    @PublicEvolving
-    static class DynamoDbSinkFailFastException extends DynamoDbSinkException {
-
-        public DynamoDbSinkFailFastException() {
-            super(
-                    "Encountered an exception while persisting records, not retrying due to {failOnError} being set.");
-        }
-
-        public DynamoDbSinkFailFastException(final Throwable cause) {
-            super(
-                    "Encountered an exception while persisting records, not retrying due to {failOnError} being set.",
-                    cause);
-        }
-    }
-}
