+++ flink-connector-dynamodb/src/test/java/org/apache/flink/connector/dynamodb/testutils/Item.java
@@ -1,0 +1,46 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.connector.dynamodb.testutils;
+
+import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
+
+import java.util.HashMap;
+import java.util.Map;
+
+/** DynamoDB item container. */
+public class Item {
+
+    public static Item.ItemBuilder builder() {
+        return new Item.ItemBuilder();
+    }
+
+    /** Builder to constrict DynamoDB item. */
+    public static final class ItemBuilder {
+        Map<String, AttributeValue> item = new HashMap<>();
+
+        public Item.ItemBuilder attr(String name, String value) {
+            item.put(name, AttributeValue.builder().s(value).build());
+            return this;
+        }
+
+        public Map<String, AttributeValue> build() {
+            return item;
+        }
+    }
+}
