+++ flink-connector-dynamodb/src/test/java/org/apache/flink/connector/dynamodb/testutils/Items.java
@@ -1,0 +1,47 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.connector.dynamodb.testutils;
+
+import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+/** Container for DynamoDB items. */
+public class Items {
+
+    public static ItemsBuilder builder() {
+        return new Items.ItemsBuilder();
+    }
+
+    /** Builder to constrict DynamoDB items. */
+    public static final class ItemsBuilder {
+        private final List<Map<String, AttributeValue>> items = new ArrayList<>();
+
+        public ItemsBuilder item(Map<String, AttributeValue> attributes) {
+            items.add(attributes);
+            return this;
+        }
+
+        public List<Map<String, AttributeValue>> build() {
+            return items;
+        }
+    }
+}
